+++ src/DatabaseDrivers/PDODatabaseDriver.php
@@ -79,6 +79,11 @@ class PDODatabaseDriver extends BaseDatabaseDriver implements DatabaseDriverInte
       return;
     }
 
+    // Don't update roles if they haven't changed
+    if ($this->getRoles($className, $foreignId) == $roles) {
+      return;
+    }
+
     $cacheKey = $this->cacheKey($className, $foreignId);
 
     $this->cache->delete($cacheKey);
